+++ engine/parsed_content.c
@@ -26,8 +26,6 @@
 
 #include <ironbee/parsed_content.h>
 
-#include <ironbee/engine.h>
-#include <ironbee/engine_types.h>
 #include <ironbee/log.h>
 #include <ironbee/mpool.h>
 
@@ -35,20 +33,20 @@
 
 ib_status_t ib_parsed_name_value_pair_list_wrapper_create(
     ib_parsed_name_value_pair_list_wrapper_t **headers,
-    ib_tx_t *tx)
+    ib_mpool_t *mp)
 {
     assert(headers != NULL);
-    assert(tx != NULL);
+    assert(mp != NULL);
 
     ib_parsed_name_value_pair_list_wrapper_t *headers_tmp =
-        ib_mpool_calloc(tx->mp, 1, sizeof(*headers_tmp));
+        ib_mpool_calloc(mp, 1, sizeof(*headers_tmp));
 
     if ( headers_tmp == NULL ) {
         *headers = NULL;
         return IB_EALLOC;
     }
 
-    headers_tmp->mpool = tx->mp;
+    headers_tmp->mpool = mp;
     /* headers_tmp->head = initialized by calloc */
     /* headers_tmp->tail = initialized by calloc */
     /* headers_tmp->size = initialized by calloc */
@@ -142,7 +140,7 @@ ib_status_t ib_parsed_tx_each_header(
 
 ib_status_t ib_parsed_resp_line_create(
     ib_parsed_resp_line_t **line,
-    ib_tx_t                *tx,
+    ib_mpool_t             *mp,
     const char             *raw,
     size_t                  raw_len,
     const char             *protocol,
@@ -155,11 +153,9 @@ ib_status_t ib_parsed_resp_line_create(
 {
     ib_status_t rc = IB_OK;
 
-    assert(tx != NULL);
-    assert(tx->ib != NULL);
-    assert(tx->mp != NULL);
+    assert(mp != NULL);
 
-    ib_parsed_resp_line_t *line_tmp = ib_mpool_alloc(tx->mp,
+    ib_parsed_resp_line_t *line_tmp = ib_mpool_alloc(mp,
                                                      sizeof(*line_tmp));
 
     if (line_tmp == NULL) {
@@ -169,7 +165,7 @@ ib_status_t ib_parsed_resp_line_create(
 
     if (protocol != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->protocol,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)protocol,
                                 protocol_len);
         if (rc != IB_OK) {
@@ -178,7 +174,7 @@ ib_status_t ib_parsed_resp_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->protocol,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -189,7 +185,7 @@ ib_status_t ib_parsed_resp_line_create(
 
     if (status != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->status,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)status,
                                 status_len);
         if (rc != IB_OK) {
@@ -198,7 +194,7 @@ ib_status_t ib_parsed_resp_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->status,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -208,7 +204,7 @@ ib_status_t ib_parsed_resp_line_create(
 
     if (msg != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->msg,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)msg,
                                 msg_len);
         if (rc != IB_OK) {
@@ -217,7 +213,7 @@ ib_status_t ib_parsed_resp_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->msg,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -232,7 +228,7 @@ ib_status_t ib_parsed_resp_line_create(
 
         if (protocol_len + status_len + msg_len == 0) {
             rc = ib_bytestr_dup_mem(&line_tmp->raw,
-                                    tx->mp,
+                                    mp,
                                     (const uint8_t *)"",
                                     0);
             if (rc != IB_OK) {
@@ -244,7 +240,7 @@ ib_status_t ib_parsed_resp_line_create(
              * the data into it.
              */
             rc = ib_bytestr_create(&line_tmp->raw,
-                                   tx->mp,
+                                   mp,
                                    protocol_len + 1 + status_len +
                                    (msg == NULL ? 0 : 1 + msg_len));
             if (rc != IB_OK) {
@@ -272,7 +268,7 @@ ib_status_t ib_parsed_resp_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->raw,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)raw,
                                 raw_len);
         if (rc != IB_OK) {
@@ -288,7 +284,7 @@ ib_status_t ib_parsed_resp_line_create(
 
 ib_status_t ib_parsed_req_line_create(
     ib_parsed_req_line_t **line,
-    ib_tx_t               *tx,
+    ib_mpool_t            *mp,
     const char            *raw,
     size_t                 raw_len,
     const char            *method,
@@ -301,11 +297,9 @@ ib_status_t ib_parsed_req_line_create(
 {
     ib_status_t rc = IB_OK;
 
-    assert(tx != NULL);
-    assert(tx->ib != NULL);
-    assert(tx->mp != NULL);
+    assert(mp != NULL);
 
-    ib_parsed_req_line_t *line_tmp = ib_mpool_alloc(tx->mp,
+    ib_parsed_req_line_t *line_tmp = ib_mpool_alloc(mp,
                                                     sizeof(*line_tmp));
 
     if ( line_tmp == NULL ) {
@@ -320,7 +314,7 @@ ib_status_t ib_parsed_req_line_create(
      */
     if (method != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->method,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)method,
                                 method_len);
         if (rc != IB_OK) {
@@ -329,7 +323,7 @@ ib_status_t ib_parsed_req_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->method,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -339,7 +333,7 @@ ib_status_t ib_parsed_req_line_create(
 
     if (uri != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->uri,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)uri,
                                 uri_len);
         if (rc != IB_OK) {
@@ -348,7 +342,7 @@ ib_status_t ib_parsed_req_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->uri,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -358,7 +352,7 @@ ib_status_t ib_parsed_req_line_create(
 
     if (protocol != NULL) {
         rc = ib_bytestr_dup_mem(&line_tmp->protocol,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)protocol,
                                 protocol_len);
         if (rc != IB_OK) {
@@ -367,7 +361,7 @@ ib_status_t ib_parsed_req_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->protocol,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)"",
                                 0);
         if (rc != IB_OK) {
@@ -378,13 +372,8 @@ ib_status_t ib_parsed_req_line_create(
     /* If no raw line is available, then create one. */
     if (raw == NULL) {
         if (method_len + uri_len + protocol_len == 0) {
-            ib_log_notice_tx(tx,
-                "Unable to generate raw request line without line "
-                "components - using zero length request line."
-            );
-
             rc = ib_bytestr_dup_mem(&line_tmp->raw,
-                                    tx->mp,
+                                    mp,
                                     (const uint8_t *)"",
                                     0);
             if (rc != IB_OK) {
@@ -402,14 +391,10 @@ ib_status_t ib_parsed_req_line_create(
              */
             raw_line_len = method_len + 1 + uri_len +
                            (protocol == NULL ? 0 : 1 + protocol_len);
-            rc = ib_bytestr_create(&line_tmp->raw, tx->mp, raw_line_len);
+            rc = ib_bytestr_create(&line_tmp->raw, mp, raw_line_len);
             if (rc != IB_OK) {
                 return rc;
             }
-
-            ib_log_debug_tx(tx,
-                            "Generating raw request line from components "
-                            "(length %zd).", raw_line_len);
 
             ib_bytestr_append_mem(line_tmp->raw,
                                   (const uint8_t *)method,
@@ -432,7 +417,7 @@ ib_status_t ib_parsed_req_line_create(
     }
     else {
         rc = ib_bytestr_dup_mem(&line_tmp->raw,
-                                tx->mp,
+                                mp,
                                 (const uint8_t *)raw,
                                 raw_len);
         if (rc != IB_OK) {